requirements.txt
flask
resuests
python-dotenv
qrcode[pil]
